Add lgwin

+++ tests/differential_c.rs
@@ -13,7 +13,7 @@ use support::{
 };
 
 /// Compares one input against the C encoder for one parameter set.
-fn assert_matches_c(name: &str, data: &[u8], lgwin: usize) {
+fn assert_matches_c(name: &str, data: &[u8], lgwin: u8) {
     let compressor = Brotli::default().compressor();
     for quality in IMPLEMENTED_QUALITIES {
         let data = prefix_for(quality, data);
@@ -49,7 +49,7 @@ fn boundary_lengths_match_the_c_encoder() {
 #[test]
 fn every_window_size_matches_the_c_encoder() {
     let corpora = structural_corpora();
-    for lgwin in 10..=24usize {
+    for lgwin in 10..=24u8 {
         for corpus in &corpora {
             if corpus.data.len() > 300_000 {
                 continue;
