+++ tests/randomized.rs
@@ -73,7 +73,7 @@ fn random_inputs_match_the_c_encoder_and_round_trip() {
     let mut rng = Rng::new(0xC0FF_EE00_1234_5678);
     for case in 0..400u32 {
         let data = generate(&mut rng, 300_000);
-        let lgwin = 10 + (rng.next_u64() as usize) % 15;
+        let lgwin = 10 + (rng.next_u64() % 15) as u8;
         for quality in IMPLEMENTED_QUALITIES {
             let data = prefix_for(quality, &data);
             let expected = c_compress(quality_number(quality), lgwin as i32, data);
@@ -100,7 +100,7 @@ fn random_inputs_agree_across_backends() {
     let mut rng = Rng::new(0x1234_5678_9ABC_DEF0);
     for case in 0..150u32 {
         let data = generate(&mut rng, 200_000);
-        let lgwin = 10 + (rng.next_u64() as usize) % 15;
+        let lgwin = 10 + (rng.next_u64() % 15) as u8;
         for quality in IMPLEMENTED_QUALITIES {
             let data = prefix_for(quality, &data);
             let mut reference: Option<Vec<u8>> = None;
@@ -127,7 +127,7 @@ fn short_random_inputs_match_the_c_encoder() {
     let mut rng = Rng::new(0x0F0F_0F0F_0F0F_0F0F);
     for case in 0..3_000u32 {
         let data = generate(&mut rng, 512);
-        let lgwin = 10 + (rng.next_u64() as usize) % 15;
+        let lgwin = 10 + (rng.next_u64() % 15) as u8;
         for quality in IMPLEMENTED_QUALITIES {
             let data = prefix_for(quality, &data);
             let expected = c_compress(quality_number(quality), lgwin as i32, data);
